마지막 예외처리 마무리(로그인, 회원가입 예외 처리)

+++ main.py
@@ -1,4 +1,5 @@
 import json
+import os.path
 from tkinter import *
 from tkinter import messagebox
 
@@ -84,19 +85,31 @@ def signWrite(pk, username, ID, password):
     user_file_path = "./json/users.json"
     user_data = {}
 
-    with open(user_file_path, "r") as json_file:
-        user_data = json.load(json_file)
+    if os.path.exists(user_file_path):
+        with open(user_file_path, "r") as json_file:
+            user_data = json.load(json_file)
 
-    user_data[pk] = {
-        "pk": pk,
-        "username": username,
-        "id": ID,
-        "password": password,
-        "balance": "0"
-    }
+        user_data[pk] = {
+            "pk": pk,
+            "username": username,
+            "id": ID,
+            "password": password,
+            "balance": "0"
+        }
 
-    with open(user_file_path, "w") as outfile:
-        json.dump(user_data, outfile, indent=4)
+        with open(user_file_path, "w") as outfile:
+            json.dump(user_data, outfile, indent=4)
+    else:
+        user_data[pk] = {
+            "pk": pk,
+            "username": username,
+            "id": ID,
+            "password": password,
+            "balance": "0"
+        }
+
+        with open(user_file_path, "w") as outfile:
+            json.dump(user_data, outfile, indent=4)
 
 
 def writeBalance(pk, balance):
@@ -112,20 +125,24 @@ def writeBalance(pk, balance):
 # 출력 함수 부분
 def read(pk):
     user_file_path = "./json/users.json"
-    with open(user_file_path, "r") as json_file:
-        json_data = json.load(json_file)
-        user = json_data.get(pk)
-        return user
+    if os.path.exists(user_file_path):
+        with open(user_file_path, "r") as json_file:
+            json_data = json.load(json_file)
+            user = json_data.get(pk)
+            return user
 
 
 def loginRead(pk):
     user_file_path = "./json/users.json"
-    with open(user_file_path, "r") as json_file:
-        json_data = json.load(json_file)
-        user = json_data.get(pk)
-        ID = user.get("id")
-        Password = user.get("password")
-        return ID, Password
+    if os.path.exists(user_file_path):
+        with open(user_file_path, "r") as json_file:
+            json_data = json.load(json_file)
+            user = json_data.get(pk)
+            ID = user.get("id")
+            Password = user.get("password")
+            return ID, Password
+    else:
+        messagebox.askyesno(title="DB 에러", message="DB가 존재하지 않습니다.")
 
 
 # 회원가입 부분
@@ -135,26 +152,45 @@ def sign_in(inputPK, inputUsername, inputID, inputPassword):
     ID = inputID.get()
     password = inputPassword.get()
     data = read(pk)
-    data_pk = data.get("pk")
-    data_user = data.get("username")
-    data_id = data.get("id")
-    data_password = data.get("password")
-    if pk == data_pk:
-        messagebox.askyesno(title="Validation Error",
-                            message="중복되는 보안 코드가 이미 있습니다.")
+    if data is not None:
+        data_pk = data.get("pk")
+        data_user = data.get("username")
+        data_id = data.get("id")
+        data_password = data.get("password")
+        if pk == data_pk:
+            messagebox.askyesno(title="Validation Error",
+                                message="중복되는 보안 코드가 이미 있습니다.")
 
-    elif username == data_user:
-        messagebox.askyesno(title="Validation Error",
-                            message="중복되는 이름이 이미 있습니다.")
-    elif ID == data_id:
-        messagebox.askyesno(title="Validation Error",
-                            message="중복되는 아이디가 이미 있습니다.")
-    elif password == data_password:
-        messagebox.askyesno(title="Validation Error",
-                            message="중복되는 비밀번호가 이미 있습니다.")
+        elif username == data_user:
+            messagebox.askyesno(title="Validation Error",
+                                message="중복되는 이름이 이미 있습니다.")
+        elif ID == data_id:
+            messagebox.askyesno(title="Validation Error",
+                                message="중복되는 아이디가 이미 있습니다.")
+        elif password == data_password:
+            messagebox.askyesno(title="Validation Error",
+                                message="중복되는 비밀번호가 이미 있습니다.")
+        else:
+            if not pk:
+                messagebox.askyesno(
+                    title="회원가입 오류", message="보안 코드를 입력하지 않았습니다.")
+            elif not username:
+                messagebox.askyesno(
+                    title="회원가입 오류", message="유저 이름을 입력하지 않았습니다.")
+            elif not ID:
+                messagebox.askyesno(
+                    title="회원가입 오류", message="아이디를 입력하지 않았습니다.")
+            elif not password:
+                messagebox.askyesno(
+                    title="회원가입 오류", message="비밀번호를 입력하지 않았습니다.")
+            else:
+                db = signWrite(pk, username, ID, password)
+                messagebox.askyesno(
+                    title="회원가입 성공", message="회원가입 성공\n '이전으로' 버튼을 눌러주십시오.")
     else:
         if not pk:
-            messagebox.askyesno(title="회원가입 오류", message="보안 코드를 입력하지 않았습니다.")
+            messagebox.askyesno(
+                title="회원가입 오류", message="보안 코드를 입력하지 않았습니다.")
         elif not username:
             messagebox.askyesno(title="회원가입 오류", message="유저 이름을 입력하지 않았습니다.")
         elif not ID:
@@ -181,33 +217,36 @@ def login(inputPK, inputID, inputPassword):
         messagebox.askyesno(title="로그인 에러", message="비밀번호를 입력하지 않았습니다.")
     else:
         db = loginRead(pk)
-        db_id = db[0]
-        db_password = db[1]
-        if db_id != ID:
-            messagebox.askyesno(
-                title="로그인 실패", message="아이디가\n 틀렸습니다.")
-        elif db_password != password:
-            messagebox.askyesno(title="로그인 실패", message="비밀번호가\n 틀렸습니다.")
+        if db is not None:
+            db_id = db[0]
+            db_password = db[1]
+            if db_id != ID:
+                messagebox.askyesno(
+                    title="로그인 실패", message="아이디가\n 틀렸습니다.")
+            elif db_password != password:
+                messagebox.askyesno(title="로그인 실패", message="비밀번호가\n 틀렸습니다.")
+            else:
+                messagebox.askyesno(title="로그인 성공", message="창을 닫고 진행하십시오.")
+                bank = BankAccount(pk)
+                while True:
+                    menu = int(
+                        input('\n원하시는 기능을 선택하세요 ( 1 : 입금, 2 : 송금, 3 : 종료 ) : '))
+                    if type(menu) != int:
+                        print("번호를 입력하십시오.")
+                    elif menu < 0 or menu > 3:
+                        print("잘못된 번호입니다. 다시 입력해주세요.")
+                    else:
+                        if menu == 1:
+                            bank.deposit(int(input('\n입금할 금액을 입력하세요(원) : ')))
+                            print('귀하의 잔액은 ', bank.get_balance(), '원 입니다.\n')
+                        elif menu == 2:
+                            bank.withdraw(int(input('\n송금할 금액을 입력하세요(원) : ')))
+                            print('귀하의 잔액은 ', bank.get_balance(), '원 입니다.\n')
+                        elif menu == 3:
+                            print('\n이용해주셔서 감사합니다.\n\n종료 버튼을 눌러주십시오.')
+                            break
         else:
-            messagebox.askyesno(title="로그인 성공", message="창을 닫고 진행하십시오.")
-            bank = BankAccount(pk)
-            while True:
-                menu = int(
-                    input('\n원하시는 기능을 선택하세요 ( 1 : 입금, 2 : 송금, 3 : 종료 ) : '))
-                if type(menu) != int:
-                    print("번호를 입력하십시오.")
-                elif menu < 0 or menu > 3:
-                    print("잘못된 번호입니다. 다시 입력해주세요.")
-                else:
-                    if menu == 1:
-                        bank.deposit(int(input('\n입금할 금액을 입력하세요(원) : ')))
-                        print('귀하의 잔액은 ', bank.get_balance(), '원 입니다.\n')
-                    elif menu == 2:
-                        bank.withdraw(int(input('\n송금할 금액을 입력하세요(원) : ')))
-                        print('귀하의 잔액은 ', bank.get_balance(), '원 입니다.\n')
-                    elif menu == 3:
-                        print('\n이용해주셔서 감사합니다.\n\n종료 버튼을 눌러주십시오.')
-                        break
+            pass
 
 
 # 은행 기능 부분
